+++ resources/views/seccion/index.blade.php
@@ -37,20 +37,26 @@
                                                 <th>operaciones</th>
                                             </tr>
                                     </tfoot>
+                                    @if(count($secciones) > 0)
                                         @foreach($secciones as $seccion)
                                         <tbody>
                                         <td>{{$seccion->id_seccion}}</td>
                                        <td>{{$seccion->nombre}}</td>
                                        <td>
+                                       @if($seccion->deleted_at =="")
                                         {!!link_to_route('seccion.edit', $title = 'Editar', $parameters = $seccion->id_seccion, $attributes = ['class'=>'btn btn-success btn-flat btn-addon m-b-10 m-l-5'])!!} 
                                     
                                         {!!Form::open(['route' => ['seccion.destroy',$seccion->id_seccion],'method'=>'DELETE'])!!} 
                                         {!!Form::submit('Eliminar',['class'=>'btn btn-inverse waves-effect waves-light'  ])!!}
                                         {!!Form::close()!!}
-
+                                         
+                                        @else
+                                        {!!link_to_route('seccion.show', $title = 'Restaurar', $parameters =  $seccion->id_seccion, $attributes = ['class'=>'btn btn-info btn-flat btn-addon m-b-10 m-l-5'] )!!}
+                                        @endif
                                        </td>
                                         </tbody>
                                        @endforeach
+                                       @endif
                                    </table>
                                 </div>
                                </div>
